+++ cmd/cove/cove_help_subsub_e2e_test.go
@@ -59,6 +59,8 @@ var subSubHelp = []subSubHelpCase{
 	{"daemon", "metrics"},
 	{"daemon", "ui"},
 	{"fleet", "add"},
+	{"fleet", "cordon"},
+	{"fleet", "uncordon"},
 	{"fleet", "ls"},
 	{"fleet", "rm"},
 	{"fleet", "vm"},